requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.4
django-registration==3.0.1
gunicorn==19.9.0
lxml==4.4.1
Pillow==6.1.0
psycopg2-binary==2.7.7
python-docx==0.8.10
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.2
